+++ src/onegov.core/onegov/core/crypto/__init__.py
@@ -1,11 +1,14 @@
 from onegov.core.crypto.password import hash_password, verify_password
 from onegov.core.crypto.random import random_password
-from onegov.core.crypto.token import random_token, RANDOM_TOKEN_LENGTH
+from onegov.core.crypto.token import random_token
+from onegov.core.crypto.token import RANDOM_TOKEN_LENGTH
+from onegov.core.crypto.token import stored_random_token
 
 __all__ = [
     'hash_password',
     'random_password',
     'random_token',
+    'stored_random_token',
     'verify_password',
     'RANDOM_TOKEN_LENGTH',
 ]